fix: Use negative rounds for fake races
Adding the 2022 calendar put all the races as previous races. This was
because the fake races started at 99 and going by count all the new
races were in the past.

In the future races should be sorted by date and not by round. A
temporary fix is to use negative rounds for fake races.